[#18] move: 파일 정리
controller 는 web 패키지의 하위 패키지로 옮겼습니다.

UserRole 은 사용하지 않기에 지웠습니다.
