--- src/main/java/sparta/gameblog/constant/UserRole.java
@@ -1,6 +1,0 @@
-package sparta.gameblog.constant;
-
-public enum UserRole {
-    NORMAL,
-    ADMIN
-}
