+++ src/main/java/sparta/gameblog/web/controller/SmtpController.java
@@ -1,4 +1,4 @@
-package sparta.gameblog.web;
+package sparta.gameblog.web.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
