+++ src/main/java/sparta/gameblog/web/controller/UserController.java
@@ -1,4 +1,4 @@
-package sparta.gameblog.web;
+package sparta.gameblog.web.controller;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
